Refactoring

+++ src/Accountancy/Features/TransactionManagement/FullfillTransaction.php
@@ -33,7 +33,7 @@ class FullfillTransaction
 
 
     /**
-     * @var double
+     * @var float
      */
     protected $amount = 0.0;
 
@@ -62,11 +62,11 @@ class FullfillTransaction
     }
 
     /**
-     * @param double $amount
+     * @param float $amount
      */
     public function setAmount($amount)
     {
-        $this->amount = (double) $amount;
+        $this->amount = (float) $amount;
     }
 
     /**
