+++ src/Accountancy/Features/TransactionManagement/IncomeTransaction.php
@@ -42,7 +42,7 @@ class IncomeTransaction
     protected $counterpartyId;
 
     /**
-     * @var double
+     * @var float
      */
     protected $amount = 0.0;
 
@@ -71,11 +71,11 @@ class IncomeTransaction
     }
 
     /**
-     * @param double $amount
+     * @param float $amount
      */
     public function setAmount($amount)
     {
-        $this->amount = (double) $amount;
+        $this->amount = (float) $amount;
     }
 
     /**
